Split tray into separate module
